Implements SSG
