+++ pages/post/[slug].js
@@ -9,6 +9,7 @@ import Icon from "supercons";
 
 import Layout from "../../components/layout";
 import HeroImage from "../../components/heroImage";
+import { getAllPosts, getPostAndMorePosts } from "../../lib/cms";
 
 const Post = ({ post, relatedPosts, mdxSource, readingTime, publishDate }) => {
   const content = hydrate(mdxSource, {});
@@ -44,48 +45,24 @@ const Post = ({ post, relatedPosts, mdxSource, readingTime, publishDate }) => {
   );
 };
 
-Post.getInitialProps = async context => {
-  const { slug } = context.query;
-  const client = createClient({
-    space: process.env.NEXT_PUBLIC_SPACE_ID,
-    accessToken: process.env.NEXT_PUBLIC_ACCESS_TOKEN
-  });
-
-  const postQuery = await client.getEntries({
-    content_type: "blogPost",
-    "fields.slug": slug
-  });
-
-  let tags;
-  let relatedPosts;
-  if (postQuery?.items[0]?.fields.hasOwnProperty(tags)) {
-    tags = postQuery.items[0].fields.tags.toString();
-    relatedPosts = await client.getEntries({
-      content_type: "blogPost",
-      select: "fields.title,fields.slug,fields.tags,fields.unsplashId",
-      "fields.tags[in]": tags,
-      "sys.id[ne]": postQuery.items[0].sys.id,
-      order: "-sys.createdAt",
-      limit: 3
-    });
-  } else {
-    tags = [];
-    relatedPosts = {
-      items: []
-    };
-  }
-
-  const post = postQuery.items[0];
-  const mdxSource = await renderToString(post.fields.body, {});
-
+export const getStaticProps = async ({ params }) => {
+  const data = await getPostAndMorePosts(params.slug);
   return {
-    post: post,
-    relatedPosts: relatedPosts.items,
-    mdxSource: mdxSource,
-    readingTime: Math.round(readingTime(post.fields.body).minutes),
-    publishDate: format(new Date(post.sys.createdAt), "dd LLL, yyyy", {
-      locale: esLocale
-    })
+    props: {
+      post: data?.post ?? null,
+      relatedPosts: data?.relatedPosts ?? null,
+      mdxSource: data?.mdxSource ?? null,
+      readingTime: data?.readingTime ?? 0,
+      publishDate: data?.publishDate
+    }
+  };
+};
+
+export const getStaticPaths = async () => {
+  const allPosts = await getAllPosts();
+  return {
+    paths: allPosts?.map(({ slug }) => `/post/${slug}`) ?? [],
+    fallback: true
   };
 };
